refactor: split inference_mixin (924->406) and pipelines (765->195+588)
inference_mixin.py decomposed into 4 focused mixins:
- inference_mixin.py: core inference control, reprocess, pipeline dispatch
- alpha_import_mixin.py: alpha import dialog + validation
- model_run_mixin.py: GVM/BiRefNet/VideoMaMa/MatAnyone2 job submission
- cancel_mixin.py: extraction/inference cancellation, force restart

pipelines.py split by annotation requirement:
- pipelines_auto.py: auto-alpha (GVM, BiRefNet) — no annotations needed
- pipelines_guided.py: guided (SAM2, VideoMaMa, MatAnyone2) — need masks

All 427 tests pass.
