+++ ui/main_window_mixins/__init__.py
@@ -3,6 +3,9 @@ from .shortcuts_mixin import ShortcutsMixin
 from .clip_mixin import ClipMixin
 from .import_mixin import ImportMixin
 from .inference_mixin import InferenceMixin
+from .alpha_import_mixin import AlphaImportMixin
+from .model_run_mixin import ModelRunMixin
+from .cancel_mixin import CancelMixin
 from .worker_mixin import WorkerMixin
 from .annotation_mixin import AnnotationMixin
 from .export_mixin import ExportMixin
@@ -11,6 +14,7 @@ from .settings_mixin import SettingsMixin
 
 __all__ = [
     "MenuMixin", "ShortcutsMixin", "ClipMixin", "ImportMixin",
-    "InferenceMixin", "WorkerMixin", "AnnotationMixin",
+    "InferenceMixin", "AlphaImportMixin", "ModelRunMixin", "CancelMixin",
+    "WorkerMixin", "AnnotationMixin",
     "ExportMixin", "SessionMixin", "SettingsMixin",
 ]
